+++ Wrappers/Python/cil/optimisation/operators/__init__.py
@@ -32,4 +32,6 @@ from .ChannelwiseOperator import ChannelwiseOperator
 from .BlurringOperator import BlurringOperator
 from .ProjectionMap import ProjectionMap
 from .WaveletOperator import WaveletOperator
-
+from .TikhonovOperator import (TikhonovOperator, BlockTikhonovOperator,
+                               WeightedStructOperator,
+                               create_tikhonov_operator, resolve_form)